requirements.txt
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.3
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.11.26
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
